+++ src/utils/calculateGrossBalance.js
@@ -1,0 +1,3 @@
+export function calculateGrossBalance(amount, targetPrice) {
+    return amount * targetPrice;
+}
