+++ src/utils/calculateProfit.js
@@ -1,0 +1,3 @@
+export function calculateProfit(grossBalance, invested) {
+    return grossBalance - invested;
+}
