+++ src/utils/isValid.js
@@ -1,0 +1,10 @@
+export function isValid(number) {
+    if (
+        isNaN(number) ||
+        number === null ||
+        number === undefined
+    ) {
+        return false;
+    }
+    return true;
+}
