+++ src/utils/parseValue.js
@@ -1,0 +1,11 @@
+import Big from 'big.js';
+
+export function parseValue(number) {
+    if (!number) return null;
+    try {
+        const value = new Big(number);
+        return value;
+    } catch (error) {
+        return null;
+    }
+}
